refactor:  Clarify the division of responsibilities for the strategy, clean up the examples output, and split up network.go
